fix: Run formatter

+++ src/Interfaces/IsTHEO1.sol
@@ -2,15 +2,15 @@
 pragma solidity 0.7.5;
 
 interface IsTHEO {
-    function rebase( uint256 ohmProfit_, uint epoch_) external returns (uint256);
+    function rebase(uint256 ohmProfit_, uint256 epoch_) external returns (uint256);
 
     function circulatingSupply() external view returns (uint256);
 
     function balanceOf(address who) external view returns (uint256);
 
-    function gonsForBalance( uint amount ) external view returns ( uint );
+    function gonsForBalance(uint256 amount) external view returns (uint256);
 
-    function balanceForGons( uint gons ) external view returns ( uint );
+    function balanceForGons(uint256 gons) external view returns (uint256);
 
-    function index() external view returns ( uint );
+    function index() external view returns (uint256);
 }
